Adjust folders + fix #37
