+++ Config/DefaultAzSpeech.ini
@@ -91,5 +91,7 @@
 +PropertyRedirects=(OldName="SSMLToSpeechAsync.SSMLToSpeech.RecognizedStream", NewName="SSMLToSpeechAsync.SSMLToSpeech.FinalAudioData")
 +PropertyRedirects=(OldName="TextToSpeechAsync.TextToSpeech.FinalAudioData", NewName="TextToSpeechAsync.TextToSpeech.bSuccess")
 +PropertyRedirects=(OldName="SSMLToSpeechAsync.SSMLToSpeech.FinalAudioData", NewName="SSMLToSpeechAsync.SSMLToSpeech.bSuccess")
++PropertyRedirects=(OldName="TextToSpeechAsync.TextToSpeech.OutputValue", NewName="TextToSpeechAsync.TextToSpeech.bSuccess")
++PropertyRedirects=(OldName="SSMLToSpeechAsync.SSMLToSpeech.OutputValue", NewName="SSMLToSpeechAsync.SSMLToSpeech.bSuccess")
 +PropertyRedirects=(OldName="SSMLToAudioDataAsync.SSMLToStream.RecognizedStream", NewName="SSMLToAudioDataAsync.SSMLToAudioData.FinalAudioData")
 +PropertyRedirects=(OldName="TextToAudioDataAsync.TextToStream.RecognizedStream", NewName="TextToAudioDataAsync.TextToAudioData.FinalAudioData")
