+++ Source/AzSpeech/Public/AzSpeech/Bases/AzSpeechAudioDataSynthesisBase.h
@@ -1,0 +1,22 @@
+// Author: Lucas Vilas-Boas
+// Year: 2022
+// Repo: https://github.com/lucoiso/UEAzSpeech
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "AzSpeech/Bases/AzSpeechSynthesizerTaskBase.h"
+#include "AzSpeechAudioDataSynthesisBase.generated.h"
+
+/**
+ *
+ */
+UCLASS(Abstract, MinimalAPI, NotPlaceable, Category = "AzSpeech", meta = (ExposedAsyncProxy = AsyncTask))
+class UAzSpeechAudioDataSynthesisBase : public UAzSpeechSynthesizerTaskBase
+{
+	GENERATED_BODY()
+
+protected:
+	virtual bool StartAzureTaskWork_Internal() override;
+	virtual void OnSynthesisUpdate(const Microsoft::CognitiveServices::Speech::SpeechSynthesisEventArgs& SynthesisEventArgs) override;
+};
