+++ Source/AzSpeech/Public/AzSpeech/Bases/AzSpeechWavFileSynthesisBase.h
@@ -5,18 +5,29 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "AzSpeech/AzSpeechSynthesizerTaskBase.h"
-#include "AzSpeechAudioDataSynthesisBase.generated.h"
+#include "AzSpeech/Bases/AzSpeechSynthesizerTaskBase.h"
+#include "AzSpeechWavFileSynthesisBase.generated.h"
 
 /**
  *
  */
 UCLASS(Abstract, MinimalAPI, NotPlaceable, Category = "AzSpeech", meta = (ExposedAsyncProxy = AsyncTask))
-class UAzSpeechAudioDataSynthesisBase : public UAzSpeechSynthesizerTaskBase
+class UAzSpeechWavFileSynthesisBase : public UAzSpeechSynthesizerTaskBase
 {
 	GENERATED_BODY()
+
+public:
+	/* Task delegate that will be called when completed */
+	UPROPERTY(BlueprintAssignable, Category = "AzSpeech")
+	FBooleanSynthesisDelegate SynthesisCompleted;
+	
+	virtual void Activate() override;
+	virtual void StopAzSpeechTask() override;
 
 protected:
 	virtual bool StartAzureTaskWork_Internal() override;
 	virtual void OnSynthesisUpdate(const Microsoft::CognitiveServices::Speech::SpeechSynthesisEventArgs& SynthesisEventArgs) override;
+
+	FString FilePath;
+	FString FileName;
 };
