+++ Source/AzSpeech/Public/AzSpeech/Tasks/SSMLToSpeechAsync.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "AzSpeech/AzSpeechAudioDataSynthesisBase.h"
+#include "AzSpeech/Bases/AzSpeechAudioDataSynthesisBase.h"
 #include "SSMLToSpeechAsync.generated.h"
 
 /**
