+++ Source/AzSpeech/Public/AzSpeech/Tasks/SSMLToWavFileAsync.h
@@ -5,32 +5,19 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "AzSpeech/AzSpeechSynthesizerTaskBase.h"
+#include "AzSpeech/Bases/AzSpeechWavFileSynthesisBase.h"
 #include "SSMLToWavFileAsync.generated.h"
 
 /**
  *
  */
 UCLASS(NotPlaceable, Category = "AzSpeech")
-class AZSPEECH_API USSMLToWavFileAsync : public UAzSpeechSynthesizerTaskBase
+class AZSPEECH_API USSMLToWavFileAsync : public UAzSpeechWavFileSynthesisBase
 {
 	GENERATED_BODY()
 
 public:
-	/* Task delegate that will be called when completed */
-	UPROPERTY(BlueprintAssignable, Category = "AzSpeech")
-	FBooleanSynthesisDelegate SynthesisCompleted;
-
 	/* Creates a Text-To-WavFile task that will convert your string to a .wav audio file */
 	UFUNCTION(BlueprintCallable, Category = "AzSpeech", meta = (BlueprintInternalUseOnly = "true", WorldContext = "WorldContextObject", DisplayName = "SSML To .wav File"))
 	static USSMLToWavFileAsync* SSMLToWavFile(const UObject* WorldContextObject, const FString& SSMLString, const FString& FilePath, const FString& FileName);
-
-	virtual void Activate() override;
-
-protected:
-	virtual bool StartAzureTaskWork_Internal() override;
-	virtual void OnSynthesisUpdate(const Microsoft::CognitiveServices::Speech::SpeechSynthesisEventArgs& SynthesisEventArgs) override;
-
-	FString FilePath;
-	FString FileName;
 };
