+++ Source/AzSpeech/Public/AzSpeech/Tasks/TextToAudioDataAsync.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "AzSpeech/AzSpeechSynthesizerTaskBase.h"
+#include "AzSpeech/Bases/AzSpeechSynthesizerTaskBase.h"
 #include "TextToAudioDataAsync.generated.h"
 
 /**
